feat: add map strains graph subcommand
