+++ bathbot/src/commands/osu/graphs/mod.rs
@@ -1,30 +1,36 @@
-use std::iter;
+use std::{iter, ops::ControlFlow};
 
-use bathbot_macros::{HasName, SlashCommand, command};
+use bathbot_macros::{HasMods, HasName, SlashCommand, command};
 use bathbot_model::{
     Countries,
     command_fields::{GameModeOption, ShowHideOption, TimezoneOption},
 };
 use bathbot_psql::model::configs::ScoreData;
-use bathbot_util::{EmbedBuilder, MessageBuilder, constants::GENERAL_ISSUE};
+use bathbot_util::{
+    EmbedBuilder, MessageBuilder,
+    constants::{GENERAL_ISSUE, OSU_BASE},
+    matcher,
+    osu::{MapIdType, ModSelection, ModsResult},
+};
 use eyre::{Report, Result, WrapErr};
 use image::{DynamicImage, GenericImageView};
 use plotters::element::{Drawable, PointCollection};
 use plotters_backend::{BackendCoord, DrawingBackend, DrawingErrorKind};
 use plotters_skia::SkiaBackend;
 use rosu_v2::{
-    prelude::{GameMode, OsuError},
+    prelude::{GameMode, GameMods, OsuError},
     request::UserId,
 };
-use score_rank::score_rank_graph;
 use time::UtcOffset;
 use twilight_interactions::command::{CommandModel, CommandOption, CreateCommand, CreateOption};
 use twilight_model::id::{Id, marker::UserMarker};
 
+pub use self::map_strains::map_strains_graph;
 use self::{
     medals::medals_graph,
     playcount_replays::{ProfileGraphFlags, playcount_replays_graph},
     rank::rank_graph,
+    score_rank::score_rank_graph,
     snipe_count::snipe_count_graph,
     sniped::sniped_graph,
     top_date::top_graph_date,
@@ -33,12 +39,17 @@ use self::{
 };
 use super::{SnipeGameMode, require_link, user_not_found};
 use crate::{
+    commands::osu::HasMods,
     core::{Context, commands::CommandOrigin},
     embeds::attachment,
-    manager::redis::osu::{CachedUser, UserArgs, UserArgsError},
+    manager::{
+        MapError,
+        redis::osu::{CachedUser, UserArgs, UserArgsError},
+    },
     util::{CachedUserExt, InteractionCommandExt, interaction::InteractionCommand},
 };
 
+mod map_strains;
 mod medals;
 mod playcount_replays;
 mod rank;
@@ -52,6 +63,8 @@ mod top_time;
 #[derive(CommandModel, CreateCommand, SlashCommand)]
 #[command(name = "graph", desc = "Display graphs about some user data")]
 pub enum Graph {
+    #[command(name = "strains")]
+    MapStrains(GraphMapStrains),
     #[command(name = "medals")]
     Medals(GraphMedals),
     #[command(name = "playcount_replays")]
@@ -66,6 +79,25 @@ pub enum Graph {
     SnipeCount(GraphSnipeCount),
     #[command(name = "top")]
     Top(GraphTop),
+}
+
+#[derive(CommandModel, CreateCommand, HasMods)]
+#[command(name = "strains", desc = "Display a map's strains over time")]
+pub struct GraphMapStrains {
+    #[command(
+        desc = "Specify a map url or map id",
+        help = "Specify a map either by map url or map id.\n\
+        If none is specified, it will search in the recent channel history \
+        and pick the first map it can find."
+    )]
+    map: Option<String>,
+    #[command(
+        desc = "Specify mods e.g. hdhr or nm",
+        help = "Specify mods either directly or through the explicit `+mods!` / `+mods` syntax e.g. `hdhr` or `+hdhr!`"
+    )]
+    mods: Option<String>,
+    #[command(desc = "Specify a gamemode")]
+    mode: Option<GameModeOption>,
 }
 
 #[derive(CommandModel, CreateCommand, HasName)]
@@ -220,6 +252,28 @@ async fn slash_graph(mut command: InteractionCommand) -> Result<()> {
 // `InteractionCommand`
 async fn graph(orig: CommandOrigin<'_>, args: Graph) -> Result<()> {
     let tuple_option = match args {
+        Graph::MapStrains(args) => {
+            let graph = match map_strains(&orig, args).await {
+                Ok(ControlFlow::Continue(bytes)) => bytes,
+                Ok(ControlFlow::Break(())) => return Ok(()),
+                Err(err) => {
+                    return Err(err.wrap_err("Failed to create map strains graph"));
+                }
+            };
+
+            let embed = EmbedBuilder::new()
+                .image(attachment("graph.png"))
+                .title(graph.title)
+                .url(graph.url);
+
+            let builder = MessageBuilder::new()
+                .embed(embed)
+                .attachment("graph.png", graph.bytes);
+
+            orig.create_message(builder).await?;
+
+            return Ok(());
+        }
         Graph::Medals(args) => {
             let user_id = match user_id!(orig, args) {
                 Some(user_id) => user_id,
@@ -388,6 +442,125 @@ async fn graph(orig: CommandOrigin<'_>, args: Graph) -> Result<()> {
 
 const W: u32 = 1350;
 const H: u32 = 711;
+
+struct MapStrains {
+    bytes: Vec<u8>,
+    title: String,
+    url: String,
+}
+
+async fn map_strains(
+    orig: &CommandOrigin<'_>,
+    args: GraphMapStrains,
+) -> Result<ControlFlow<(), MapStrains>> {
+    let mods_res = args.mods();
+
+    let map = match args.map.map(|arg| {
+        matcher::get_osu_map_id(&arg)
+            .map(MapIdType::Map)
+            .or_else(|| matcher::get_osu_mapset_id(&arg).map(MapIdType::Set))
+    }) {
+        Some(Some(id)) => Some(id),
+        Some(None) => {
+            let content =
+                "Failed to parse map url. Be sure you specify a valid map id or url to a map.";
+
+            return orig.error(content).await.map(ControlFlow::Break);
+        }
+        None => None,
+    };
+
+    let map_id = if let Some(id) = map {
+        id
+    } else {
+        let Ok(msgs) = Context::retrieve_channel_history(orig.channel_id()).await else {
+            let content = "No beatmap specified and lacking permission to search the channel \
+            history for maps.\nTry specifying a map either by url or by map id, or give me the \
+            \"Read Message History\" permission.";
+
+            return orig.error(content).await.map(ControlFlow::Break);
+        };
+
+        match Context::find_map_id_in_msgs(&msgs, 0).await {
+            Some(id) => id,
+            None => {
+                let content = "No beatmap specified and none found in recent channel history. Try \
+                specifying a map either by url or by map id.";
+
+                return orig.error(content).await.map(ControlFlow::Break);
+            }
+        }
+    };
+
+    let MapIdType::Map(map_id) = map_id else {
+        let content = "Looks like you gave me a mapset id, I need a map id though";
+
+        return orig.error(content).await.map(ControlFlow::Break);
+    };
+
+    let mode = args.mode.map(GameMode::from);
+
+    let map = match Context::osu_map().map(map_id, None).await {
+        Ok(mut map) => {
+            if let Some(mode) = mode {
+                map.convert_mut(mode);
+            }
+
+            map
+        }
+        Err(MapError::NotFound) => {
+            let content = format!(
+                "Could not find beatmap with id `{map_id}`. \
+                        Did you give me a mapset id instead of a map id?",
+            );
+
+            return orig.error(content).await.map(ControlFlow::Break);
+        }
+        Err(MapError::Report(err)) => {
+            let _ = orig.error(GENERAL_ISSUE).await;
+
+            return Err(err);
+        }
+    };
+
+    let mode = mode.unwrap_or(map.mode());
+
+    let mods = match mods_res {
+        ModsResult::Mods(ModSelection::Include(mods) | ModSelection::Exact(mods)) => {
+            match mods.try_with_mode(mode) {
+                Some(mods) if mods.is_valid() => mods,
+                Some(_) => {
+                    let content = format!(
+                        "Looks like some mods in `{mods}` are incompatible with each other"
+                    );
+
+                    return orig.error(content).await.map(ControlFlow::Break);
+                }
+                None => {
+                    let content =
+                        format!("The mods `{mods}` are incompatible with the mode {mode:?}");
+
+                    return orig.error(content).await.map(ControlFlow::Break);
+                }
+            }
+        }
+        ModsResult::Mods(ModSelection::Exclude { .. }) | ModsResult::None => GameMods::new(),
+        ModsResult::Invalid => {
+            let content = "Failed to parse mods.\n\
+            If you want included mods, specify it e.g. as `+hrdt`.\n\
+            If you want exact mods, specify it e.g. as `+hdhr!`.\n\
+            And if you want to exclude mods, specify it e.g. as `-hdnf!`.";
+
+            return orig.error(content).await.map(ControlFlow::Break);
+        }
+    };
+
+    let bytes = map_strains_graph(&map.pp_map, mods, map.cover(), W, H).await?;
+    let title = format!("{} - {} [{}]", map.artist(), map.title(), map.version());
+    let url = format!("{OSU_BASE}b/{}", map.map_id());
+
+    Ok(ControlFlow::Continue(MapStrains { bytes, title, url }))
+}
 
 async fn top_graph(
     orig: &CommandOrigin<'_>,
